agent/pyproject.toml
[tool.poetry]
name = "agent"
version = "0.1.0"
description = ""
authors = ["CopilotKit"]
readme = "README.md"
packages = [{include = "sample_agent"}]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
copilotkit = "0.1.72"
langchain = "^1.0.0"
langchain-google-genai = "^2.0.0"
langgraph = "^1.0.0"
langsmith = "0.4.4"
fastapi = "^0.115.5"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.0"
langchain-ollama = "^1.0.1"
huggingface-hub = ">=0.33.4,<1.0.0"
langchain-huggingface = "^1.2.0"
aiohttp = "^3.13.2"

[tool.poetry.scripts]
demo = "sample_agent.demo:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
